[COMMON]mmc: dw_mmc: fix prevent issue

+++ drivers/mmc/host/dw_mmc.h
@@ -942,8 +942,8 @@ struct dw_mci_req_log {
 	u32 info1;
 	u32 info2;
 	u32 info3;
-	u32 pending_events;
-	u32 completed_events;
+	unsigned long pending_events;
+	unsigned long completed_events;
 	enum dw_mci_state state;
 	enum dw_mci_state state_cmd;
 	enum dw_mci_state state_dat;
